+++ panther/_utils.py
@@ -1,4 +1,5 @@
 import importlib
+import logging
 import re
 from collections.abc import Callable
 from traceback import TracebackException
@@ -8,7 +9,9 @@ import orjson as json
 
 from panther import status
 from panther.file_handler import File
-from panther.logger import logger
+
+
+logger = logging.getLogger('panther')
 
 
 async def _http_response_start(send: Callable, /, headers: dict, status_code: int) -> None:
@@ -42,8 +45,7 @@ async def http_response(
     elif status_code == status.HTTP_204_NO_CONTENT or body == b'null':
         body = None
 
-    if monitoring is not None:
-        await monitoring.after(status_code=status_code)
+    await monitoring.after(status_code=status_code)
 
     await _http_response_start(send, headers=headers, status_code=status_code)
     await _http_response_body(send, body=body)