+++ panther/middlewares/redis.py
@@ -1,11 +1,14 @@
+import logging
 from redis import Redis
 
 from panther.db.connection import RedisConnection
-from panther.logger import logger
 from panther.middlewares.base import BaseMiddleware
 from panther.request import Request
 from panther.response import Response
 from panther.websocket import GenericWebsocket
+
+
+logger = logging.getLogger('panther')
 
 
 class RedisMiddleware(BaseMiddleware):
@@ -35,4 +38,6 @@ class RedisMiddleware(BaseMiddleware):
         return response
 
     def redis_connection_for_ws(self) -> Redis:
-        return Redis(**self.kwargs)
+        r = Redis(**self.kwargs)
+        r.ping()
+        return r