+++ panther/utils.py
@@ -1,9 +1,12 @@
+import logging
+import os
 import hashlib
 from datetime import datetime, timedelta
 from pathlib import Path
 from typing import ClassVar
 
-from panther.logger import logger
+
+logger = logging.getLogger('panther')
 
 
 class Singleton(object):
@@ -30,6 +33,9 @@ def load_env(env_file: str | Path, /) -> dict[str, str]:
                 key = key.strip()
                 value = value.strip().strip('"\'')
                 variables[key] = value
+
+                # Load them as system environment variable
+                os.environ[key] = value
     return variables